Redo FetchTransformSaveApp and queueing
* flattens FetchTransformSaveApp classes into a single class
* removes unused functionality
* renames queue-related things from "new_crash_source" to "queue"
* breaks queue consuming things out of crashstorage things
