+++ socorro/external/rabbitmq/crashstorage.py
@@ -7,28 +7,13 @@ from configman import Namespace
 from configman.converters import class_converter
 from configman.dotdict import DotDict
 import pika
-from queue import Queue, Empty
 
 from socorro.external.crashstorage_base import CrashStorageBase
 from socorro.lib.transaction import retry
 
 
 class RabbitMQCrashStorage(CrashStorageBase):
-    """This class is an implementation of a Socorro Crash Storage system.
-    It is used as a crash queing methanism for raw crashes.  It implements
-    the save_raw_crash method as a queue submission function, and the
-    new_crashes generator as a queue consumption function.  Please note: as
-    it only queues the crash_id and not the whole raw crash, it is not suitable
-    to actually save a crash.  It is a very lossy container.  This class
-    should be used in conjuction with a more persistant storage mechanism.
-
-    The implementations CrashStorage classes can use arbitrarly high or low
-    level semantics to talk to their underlying resource.
-
-    The 'new_crashes' generator has a lower level relationship with the
-    underlying connection object
-
-    """
+    """Save crash ids to the specified routing key."""
 
     required_config = Namespace()
     required_config.add_option(
@@ -55,12 +40,6 @@ class RabbitMQCrashStorage(CrashStorageBase):
 
     def __init__(self, config, namespace='', quit_check_callback=None):
         super().__init__(config, namespace=namespace, quit_check_callback=quit_check_callback)
-
-        # Note: this may continue to grow if we aren't acking certain UUIDs.
-        # We should find a way to time out UUIDs after a certain time.
-        self.acknowledgement_token_cache = {}
-        self.acknowledgment_queue = Queue()
-
         self.rabbitmq = config.rabbitmq_class(config)
         self._basic_properties = pika.BasicProperties(
             # Make message persistent
@@ -101,120 +80,6 @@ class RabbitMQCrashStorage(CrashStorageBase):
             routing_key=self.config.routing_key,
             body=crash_id,
             properties=self._basic_properties
-        )
-
-    def _basic_get(self, conn, queue):
-        return conn.channel.basic_get(queue=queue)
-
-    def new_crashes(self):
-        """This generator fetches crash_ids from RabbitMQ."""
-
-        # We've set up RabbitMQ to require acknowledgement of processing of a
-        # crash_id from this generator.  It is the responsibility of the
-        # consumer of the crash_id to tell this instance of the class when has
-        # completed its work on the crash_id.  That is done with the call to
-        # 'ack_crash' below.  Because RabbitMQ connections are not thread safe,
-        # only the thread that read the crash may acknowledge it.  'ack_crash'
-        # queues the crash_id. The '_consume_acknowledgement_queue' function
-        # is run to send acknowledgments back to RabbitMQ
-        self._consume_acknowledgement_queue()
-        queues = [
-            self.rabbitmq.config.priority_queue_name,
-            self.rabbitmq.config.standard_queue_name,
-            self.rabbitmq.config.reprocessing_queue_name,
-            self.rabbitmq.config.priority_queue_name,
-        ]
-        while True:
-            for queue in queues:
-                method_frame, header_frame, body = retry(
-                    self.rabbitmq,
-                    self.quit_check,
-                    self._basic_get,
-                    queue=queue
-                )
-                # The body is always a string, so convert it to a string
-                if body:
-                    body = body.decode('utf-8')
-
-                if method_frame and self._suppress_duplicate_jobs(body, method_frame):
-                    continue
-                if method_frame:
-                    break
-            # must consume ack queue before testing for end of iterator
-            # or the last job won't get ack'd
-            self._consume_acknowledgement_queue()
-            if not method_frame:
-                # there was nothing in the queue - leave the iterator
-                return
-            self.acknowledgement_token_cache[body] = method_frame
-            yield body
-            queues.reverse()
-
-    def ack_crash(self, crash_id):
-        self.acknowledgment_queue.put(crash_id)
-
-    def _suppress_duplicate_jobs(self, crash_id, acknowledgement_token):
-        """if this crash is in the cache, then it is already in progress
-        and this is a duplicate.  Acknowledge it, then return to True
-        to let the caller know to skip on to the next crash."""
-        if crash_id in self.acknowledgement_token_cache:
-            # reject this crash - it's already being processsed
-            self.logger.info('duplicate job: %s is already in progress', crash_id)
-            # ack this
-            retry(
-                self.rabbitmq,
-                self.quit_check,
-                self._ack_crash,
-                crash_id=crash_id,
-                acknowledgement_token=acknowledgement_token
-            )
-            return True
-        return False
-
-    def _consume_acknowledgement_queue(self):
-        """The acknowledgement of the processing of each crash_id yielded
-        from the 'new_crashes' method must take place on the same connection
-        that the crash_id came from.  The crash_ids are queued in the
-        'acknowledgment_queue'.  That queue is consumed by the QueuingThread"""
-        try:
-            while True:
-                crash_id_to_be_acknowledged = self.acknowledgment_queue.get_nowait()
-
-                try:
-                    acknowledgement_token = self.acknowledgement_token_cache[
-                        crash_id_to_be_acknowledged
-                    ]
-                    retry(
-                        self.rabbitmq,
-                        self.quit_check,
-                        self._ack_crash,
-                        crash_id=crash_id_to_be_acknowledged,
-                        acknowledgement_token=acknowledgement_token
-                    )
-                    del self.acknowledgement_token_cache[crash_id_to_be_acknowledged]
-                except KeyError:
-                    self.logger.warning(
-                        'RabbitMQCrashStorage tried to acknowledge crash %s'
-                        ', which was not in the cache',
-                        crash_id_to_be_acknowledged,
-                        exc_info=True
-                    )
-                except Exception:
-                    self.logger.error(
-                        'RabbitMQCrashStorage unexpected failure on %s',
-                        crash_id_to_be_acknowledged,
-                        exc_info=True
-                    )
-
-        except Empty:
-            pass  # nothing to do with an empty queue
-
-    def _ack_crash(self, connection, crash_id, acknowledgement_token):
-        connection.channel.basic_ack(delivery_tag=acknowledgement_token.delivery_tag)
-        self.logger.debug(
-            'RabbitMQCrashStorage acking %s with delivery_tag %s',
-            crash_id,
-            acknowledgement_token.delivery_tag
         )
