+++ socorro/unittest/app/test_fetch_transform_save_app.py
@@ -22,9 +22,8 @@ class TestFetchTransformSaveApp(object):
             def _setup_source_and_destination(self):
                 self.source = Mock()
                 self.destination = Mock()
-                pass
 
-            def _create_iter(self):
+            def get_crashids(self):
                 for x in range(5):
                     yield ((x,), {})
 
@@ -35,14 +34,15 @@ class TestFetchTransformSaveApp(object):
             'logger': MagicMock(),
             'number_of_threads': 2,
             'maximum_queue_size': 2,
-            'number_of_submissions': 'all',
+            'queue': DotDict({'crashqueue_class': None}),
             'source': DotDict({'crashstorage_class': None}),
             'destination': DotDict({'crashstorage_class': None}),
             'producer_consumer': DotDict({
                 'producer_consumer_class': TaskManager,
                 'logger': MagicMock(),
                 'number_of_threads': 1,
-                'maximum_queue_size': 1
+                'maximum_queue_size': 1,
+                'quit_on_empty_queue': True,
             })
         })
 
@@ -53,9 +53,20 @@ class TestFetchTransformSaveApp(object):
 
     def test_bogus_source_and_destination(self):
         class NonInfiniteFTSAppClass(FetchTransformSaveApp):
-            def _basic_iterator(self):
+            def _infinite_iterator(self):
+                # NOTE(willkg): This isn't in an infinite loop, so it ends
                 for x in self.source.new_crashes():
                     yield ((x,), {})
+
+        class FakeQueue(object):
+            def __init__(self, config, namespace='', quit_check_callback=None):
+                pass
+
+            def new_crashes(self):
+                # NOTE(willkg): these are keys in the FakeStorageSource
+                crash_ids = ['1234', '1235', '1236', '1237']
+                for crash_id in crash_ids:
+                    yield crash_id
 
         class FakeStorageSource(object):
             def __init__(self, config, namespace='', quit_check_callback=None):
@@ -113,7 +124,7 @@ class TestFetchTransformSaveApp(object):
             'logger': MagicMock(),
             'number_of_threads': 2,
             'maximum_queue_size': 2,
-            'number_of_submissions': 'all',
+            'queue': DotDict({'crashqueue_class': FakeQueue}),
             'source': DotDict({'crashstorage_class': FakeStorageSource}),
             'destination': DotDict({'crashstorage_class': FakeStorageDestination}),
             'producer_consumer': DotDict({
@@ -137,11 +148,10 @@ class TestFetchTransformSaveApp(object):
         assert source.number_of_close_calls == 1
         assert destination.number_of_close_calls == 1
 
-    def test_source_iterator(self):
+    def test_queue_iterator(self):
         faked_finished_func = Mock()
 
-        class FakeStorageSource(object):
-
+        class FakeQueue(object):
             def __init__(self):
                 self.first = True
 
@@ -177,8 +187,8 @@ class TestFetchTransformSaveApp(object):
             'logger': MagicMock(),
             'number_of_threads': 2,
             'maximum_queue_size': 2,
-            'number_of_submissions': 'forever',
-            'source': DotDict({'crashstorage_class': FakeStorageSource}),
+            'queue': DotDict({'crashqueue_class': FakeQueue}),
+            'source': DotDict({'crashstorage_class': None}),
             'destination': DotDict({'crashstorage_class': FakeStorageDestination}),
             'producer_consumer': DotDict({
                 'producer_consumer_class': ThreadedTaskManager,
@@ -189,7 +199,8 @@ class TestFetchTransformSaveApp(object):
         })
 
         fts_app = FetchTransformSaveApp(config)
-        fts_app.source = FakeStorageSource()
+        fts_app.queue = FakeQueue()
+        fts_app.source = None
         fts_app.destination = FakeStorageDestination
         error_detected = False
         no_finished_function_counter = 0
@@ -228,7 +239,7 @@ class TestFetchTransformSaveApp(object):
             'logger': MagicMock(),
             'number_of_threads': 2,
             'maximum_queue_size': 2,
-            'number_of_submissions': 'forever',
+            'queue': DotDict({'crashqueue_class': None}),
             'source': DotDict({'crashstorage_class': None}),
             'destination': DotDict({'crashstorage_class': FakeStorageDestination}),
             'producer_consumer': DotDict({
@@ -284,7 +295,7 @@ class TestFetchTransformSaveApp(object):
             'logger': MagicMock(),
             'number_of_threads': 2,
             'maximum_queue_size': 2,
-            'number_of_submissions': 'forever',
+            'queue': DotDict({'crashqueue_class': None}),
             'source': DotDict({'crashstorage_class': FakeStorageSource}),
             'destination': DotDict({'crashstorage_class': None}),
             'producer_consumer': DotDict({
