+++ socorro/unittest/external/rabbitmq/test_crashstorage.py
@@ -111,19 +111,7 @@ class TestCrashStorage(object):
                 crash_id='crash_id'
             )
 
-    def test__save_raw_crash_normal(self):
-        connection = Mock()
-        config = self._setup_config()
-        crash_store = RabbitMQCrashStorage(config)
-        crash_store._save_raw_crash(connection, 'some_crash_id')
-        connection.channel.basic_publish.assert_called_once_with(
-            exchange='',
-            routing_key='socorro.normal',
-            body='some_crash_id',
-            properties=crash_store._basic_properties
-        )
-
-    def test__save_raw_crash_priority(self):
+    def test_save_raw_crash_priority(self):
         connection = Mock()
         config = self._setup_config()
         config.routing_key = 'socorro.priority'
@@ -135,74 +123,3 @@ class TestCrashStorage(object):
             body='some_crash_id',
             properties=crash_store._basic_properties
         )
-
-    def test__ack_crash(self):
-        config = self._setup_config()
-        connection = Mock()
-        ack_token = DotDict()
-        ack_token.delivery_tag = 1
-        crash_id = 'some-crash-id'
-
-        crash_store = RabbitMQCrashStorage(config)
-        crash_store._ack_crash(connection, crash_id, ack_token)
-
-        connection.channel.basic_ack.assert_called_once_with(delivery_tag=1)
-
-    def test_ack_crash_fails_gracefully(self, caplogpp):
-        caplogpp.set_level('WARNING')
-
-        config = self._setup_config()
-        crash_store = RabbitMQCrashStorage(config)
-        crash_store.acknowledgment_queue.put('b2')
-        crash_store._consume_acknowledgement_queue()
-
-        recs = [(rec.message, rec.exc_info) for rec in caplogpp.records]
-        assert (
-            recs[0][0] == (
-                'RabbitMQCrashStorage tried to acknowledge crash b2, which was not in the cache'
-            )
-        )
-        assert recs[0][1] is not None
-
-    def test_ack_crash(self):
-        config = self._setup_config()
-        crash_store = RabbitMQCrashStorage(config)
-        crash_store.acknowledgment_queue = Mock()
-
-        crash_store.ack_crash('crash_id')
-
-        crash_store.acknowledgment_queue.put.assert_called_once_with(
-            'crash_id'
-        )
-
-    def test_new_crash_standard_queue(self):
-        """ Tests queue with standard queue items only
-        """
-        config = self._setup_config()
-        crash_store = RabbitMQCrashStorage(config)
-        crash_store.rabbitmq.config.standard_queue_name = 'socorro.normal'
-        crash_store.rabbitmq.config.reprocessing_queue_name = 'socorro.reprocessing'
-        crash_store.rabbitmq.config.priority_queue_name = 'socorro.priority'
-
-        test_queue = [
-            ('1', '1', 'normal_crash_id'),
-            (None, None, None),
-            (None, None, None),
-        ]
-
-        def basic_get(queue):
-            if len(test_queue) == 0:
-                return
-            if queue == 'socorro.priority':
-                return (None, None, None)
-            elif queue == 'socorro.reprocessing':
-                return (None, None, None)
-            elif queue == 'socorro.normal':
-                return test_queue.pop()
-
-        crash_store.rabbitmq.return_value.__enter__.return_value  \
-            .channel.basic_get = MagicMock(side_effect=basic_get)
-
-        expected = ['normal_crash_id']
-        for result in crash_store.new_crashes():
-            assert expected.pop() == result
